fix(config): describe manifests, not only package.json
`packageRepositoryUrl({ require: true })` blamed `package.json` for a
missing or unrecognised `repository` even when the data came from
`deno.json` or `jsr.json`; the error now says manifest. The module
header, type docs, and examples of `package-json.ts` describe the
generic walk-up and import from `@kjanat/dreamcli/config`, the
repository URL tests live beside their implementation, and the config
and json-schema maintainer notes match the current file layout and
load paths.
